chore: workaround nuxt module builder
