+++ playground/layers/vuetify-layer/nuxt.config.ts
@@ -1,6 +1,5 @@
 export default defineNuxtConfig({
   // since it is local, the path is relative to the playground folder
-  modules: ['../src/module'],
   // vuetify: {
   // vuetifyOptions: './vuetify.config.ts',
   /* vuetifyOptions: {
