+++ src/module.ts
@@ -7,8 +7,15 @@ import {
   useLogger,
 } from '@nuxt/kit'
 import { getPackageInfo } from 'local-pkg'
+import type { HookResult } from '@nuxt/schema'
+import type { VuetifyOptions, createVuetify } from 'vuetify'
 import { version } from '../package.json'
-import type { ModuleOptions } from './types'
+import type {
+  InlineModuleOptions,
+  SSRClientHints,
+  SSRClientHintsConfiguration,
+  VuetifyModuleOptions,
+} from './types'
 import type { VuetifyNuxtContext } from './utils/config'
 import { load, registerWatcher } from './utils/loader'
 import { configureVite } from './utils/configure-vite'
@@ -19,7 +26,7 @@ export * from './types'
 const CONFIG_KEY = 'vuetify'
 const logger = useLogger(`nuxt:${CONFIG_KEY}`)
 
-export default defineNuxtModule<ModuleOptions>({
+export default defineNuxtModule<VuetifyModuleOptions>({
   meta: {
     name: 'vuetify-nuxt-module',
     configKey: 'vuetify',
@@ -77,3 +84,38 @@ export default defineNuxtModule<ModuleOptions>({
     configureVite(CONFIG_KEY, nuxt, ctx)
   },
 })
+
+export interface ModuleOptions extends VuetifyModuleOptions {}
+
+export interface ModuleHooks {
+  'vuetify:registerModule': (registerModule: (config: InlineModuleOptions) => void) => HookResult
+}
+
+// rename this to ModuleRuntimeHooks when released:
+// https://github.com/nuxt/module-builder/pull/194
+export interface RuntimeModuleHooks {
+  'vuetify:configuration': (options: {
+    isDev: boolean
+    vuetifyOptions: VuetifyOptions
+  }) => HookResult
+  'vuetify:before-create': (options: {
+    isDev: boolean
+    vuetifyOptions: VuetifyOptions
+  }) => HookResult
+  'vuetify:ready': (vuetify: ReturnType<typeof createVuetify>) => HookResult
+  'vuetify:ssr-client-hints': (options: {
+    vuetifyOptions: VuetifyOptions
+    ssrClientHints: SSRClientHints
+    ssrClientHintsConfiguration: SSRClientHintsConfiguration
+  }) => HookResult
+}
+
+declare module '#app' {
+  interface RuntimeNuxtHooks extends RuntimeModuleHooks {}
+}
+
+declare module '@nuxt/schema' {
+  interface NuxtConfig { ['vuetify']?: Partial<ModuleOptions> }
+  interface NuxtOptions { ['vuetify']?: ModuleOptions }
+  interface NuxtHooks extends ModuleHooks {}
+}
