+++ src/runtime/plugins/client-hints.ts
@@ -1,25 +1,1 @@
-import type { UnwrapNestedRefs } from 'vue'
-
-export interface ClientHintRequestFeatures {
-  firstRequest: boolean
-  prefersColorSchemeAvailable: boolean
-  prefersReducedMotionAvailable: boolean
-  viewportHeightAvailable: boolean
-  viewportWidthAvailable: boolean
-}
-export interface SSRClientHints extends ClientHintRequestFeatures {
-  prefersColorScheme?: 'dark' | 'light' | 'no-preference'
-  prefersReducedMotion?: 'no-preference' | 'reduce'
-  viewportHeight?: number
-  viewportWidth?: number
-  colorSchemeFromCookie?: string
-  colorSchemeCookie?: string
-}
-
 export const VuetifyHTTPClientHints = 'vuetify:nuxt:ssr-client-hints'
-
-declare module '#app' {
-  interface NuxtApp {
-    $ssrClientHints: UnwrapNestedRefs<SSRClientHints>
-  }
-}
