+++ src/runtime/plugins/vuetify-no-client-hints.ts
@@ -1,6 +1,6 @@
 import type { UnwrapNestedRefs } from 'vue'
 import { reactive } from 'vue'
-import type { SSRClientHints } from './client-hints'
+import type { SSRClientHints } from './types'
 import type { Plugin } from '#app'
 import { defineNuxtPlugin } from '#imports'
